webapi-v1/src/pyproject.toml
[tool.poetry]
name = "psa-webapi"
version = "0.1.0"
description = ""
authors = ["Slawomir Siudek <siudeks@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
graphene_django = "^2.2"
django-graphql-jwt = "^0.2.1"
django-cors-headers = "^3.0"
psycopg2 = "^2.8"

[tool.poetry.dev-dependencies]
pylint-django = "^2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
